refactor: Update CardModulo component to use Next.js Link component for "Saiba mais" button
This commit updates the CardModulo component to use the Next.js Link component instead of the <button> tag for the "Saiba mais" button. The Link component provides better performance and optimization for navigation within the application. Additionally, the button is wrapped in a Link component with the "/produtos" path specified as the href attribute. This change improves the overall efficiency and user experience of the component.